refactor json len operator
